pyproject.toml
[project]
name = "ascii-art"
version = "0.1.0"
description = "A Python library for generating ASCII art from images with advanced filtering capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pillow>=8.0.0",
    "scipy>=1.7.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.scripts]
ascii-art = "ascii_art.cli:main"

[project.urls]
Homepage = "https://github.com/adhipkashyap/ascii-art"
Repository = "https://github.com/adhipkashyap/ascii-art"
Documentation = "https://github.com/adhipkashyap/ascii-art#readme"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
